More efficient required build detection.
Rather than using "build_rels" which don't really exist in git define:

* A Dockerfile in a directory with SCM changes needs building.
* A Dockerfile with dependants that need building needs building.

Thus, the most recent commit of a Dockerfile directory and the directories of it's parents defines the most recent commit that needs building.

This is simpler, does not need to grab all the commits in the repo, and crucially **allows pulling images before building them** in a future change.